Split predict_pty_helper into three cross-platform binaries
Replace the single unix-only predict-pty-helper binary with three
focused, cross-platform binaries:

- pty-echo: byte-by-byte stdin→stdout echo loop
- pty-password: password prompt via rpassword + echo loop
- pty-interactive: mini-shell with passwd command and Ctrl+C handling

Uses rpassword (cross-platform) and ctrlc crate instead of raw libc
signal handling, enabling Windows support.
